requirements.txt
pyyaml>=6.0
requests>=2.31.0
playwright>=1.40.0
networkx>=3.0
anthropic>=0.40.0
matplotlib>=3.8.0
pytest>=7.0.0
